requirements.txt
python-binance==1.0.28
numpy>=1.20.0
pandas>=1.3.0
ta>=0.10.0
python-dotenv>=0.19.0
schedule>=1.1.0
websocket-client>=1.2.1
matplotlib>=3.5.0
ccxt>=2.0.0
requests>=2.26.0
tqdm>=4.62.0